Make everything f32 that was f64.

+++ src/game.rs
@@ -44,7 +44,7 @@ pub trait Game: Clone + Debug + Display {
     fn get_victory_state(&self) -> VictoryState;
     fn get_player(&self) -> Player;
 
-    fn exploration_factor(&self) -> f64 {
+    fn exploration_factor(&self) -> f32 {
         1.0
     }
 }
